Add support for luks-encrypted volumes
This includes two new storage classes, cloudscale-volume-ssd-luks and
cloudscale-volume-bulk-luks that encrypt the disk with aes-xts-plain64.

For disk encryption with the new storage classes, a secret with the name
${pvc.name}-luks-key must be created in the pvc's namespace. The secret
must contain the luks key in an element called luksKey.

To ease integration testing, a shell-script was added to the Dockerfile
that can be used to inspect mounted persistent volumes on a node.